increasing the on-death sleep interval if it dies immediately
the important metric is the maximum "oversleep", e.g. the value
of the sleep counter (in minutes) reached after a downtime of n minutes,
which is the inverse of the triangular number function, and thus
grows ever slower.

On first message not containing the "cannot make IMAP client" string,
the counter is reset to 1, which is its initial value.